refactor(tests): split test_dispatch.py to meet LOC limit
- Extract TestDispatchNotLaunched to test_dispatch_not_launched.py
- Original file: 379 lines (was 431, limit 400)
- New file: 78 lines
- All 7 tests pass

Fixes CI BLOCK: tests/vibe3/domain/handlers/test_dispatch.py has 431 lines
